fix float32 conversion issue

+++ cleverhans/utils.py
@@ -258,7 +258,7 @@ def get_logits_over_interval(sess, model, x, adv_x, x_data,
     channels = shape[3]
     size = height * width * channels
 
-    epsilon_array_t = tf.constant(epsilon_array)
+    epsilon_array_t = tf.constant(epsilon_array, dtype=tf.float32)
     epsilon_array_t = tf.stack([epsilon_array_t] * size, axis=1)
     epsilon_array_t = tf.reshape(epsilon_array_t,
                                  shape=[num_points, height,
